+++ middlewarehouse/api/payloads/quantity_level.go
@@ -1,0 +1,7 @@
+package payloads
+
+// QuantityLevel is an optional numerical level.
+type QuantityLevel struct {
+	IsEnabled bool `json:"isEnabled"`
+	Level     int  `json:"level"`
+}
